requirements.txt
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.24.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain==0.2.11
langchain-core==0.2.24
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.2
langsmith==0.1.93
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
openlm==0.0.5
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==4.1.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.3
setuptools==72.0.0
six==1.16.0
SQLAlchemy==2.0.31
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@f7396876849926afa87c9412d67c43618dad403d
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
Werkzeug==3.0.3
yarl==1.9.4
